api calls currencies and rates business logic. Viewmodel implementation for currency conversion. Optimize imports Reformat code

+++ app/src/main/java/com/timife/cowryconverter/data/models/RatesResponse.kt
@@ -12,5 +12,5 @@ data class RatesResponse(
     @SerializedName("date")
     val date: String?,
     @SerializedName("rates")
-    val rates:Map<String, Long>?,
+    val rates: Map<String, Double>?,
 )
