+++ app/src/main/java/com/timife/cowryconverter/domain/model/Rate.kt
@@ -2,5 +2,5 @@ package com.timife.cowryconverter.domain.model
 
 data class Rate(
     val currency: String,
-    val rate: Long
+    val rate: Double
 )
